[common] Move ErrorUi to common module and use it android
