+++ android/src/main/java/dev/sergiobelda/todometer/ui/home/HomeUiState.kt
@@ -18,6 +18,7 @@ package dev.sergiobelda.todometer.ui.home
 
 import dev.sergiobelda.todometer.common.domain.model.Task
 import dev.sergiobelda.todometer.common.domain.model.TaskList
+import dev.sergiobelda.todometer.common.ui.error.ErrorUi
 
 data class HomeUiState(
     val isLoadingTasks: Boolean = false,
@@ -25,5 +26,5 @@ data class HomeUiState(
     val taskLists: List<TaskList> = emptyList(),
     val taskListSelected: TaskList? = null,
     val isDefaultTaskListSelected: Boolean = true,
-    val errorMessage: String? = null
+    val errorUi: ErrorUi? = null
 )
