+++ android/src/main/java/dev/sergiobelda/todometer/ui/taskdetail/TaskDetailViewModel.kt
@@ -24,6 +24,7 @@ import androidx.lifecycle.viewModelScope
 import dev.sergiobelda.todometer.common.domain.doIfError
 import dev.sergiobelda.todometer.common.domain.doIfSuccess
 import dev.sergiobelda.todometer.common.domain.usecase.GetTaskUseCase
+import dev.sergiobelda.todometer.common.ui.error.mapToErrorUi
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
@@ -42,13 +43,13 @@ class TaskDetailViewModel(
                     taskDetailUiState = taskDetailUiState.copy(
                         isLoading = false,
                         task = task,
-                        errorMessage = null
+                        errorUi = null
                     )
                 }.doIfError { error ->
                     taskDetailUiState = taskDetailUiState.copy(
                         isLoading = false,
                         task = null,
-                        errorMessage = error.message
+                        errorUi = error.mapToErrorUi()
                     )
                 }
             }
