+++ common/src/commonMain/kotlin/dev/sergiobelda/todometer/common/ui/error/ErrorUi.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Sergio Belda
+ * Copyright 2022 Sergio Belda
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.sergiobelda.todometer.ui.error
+package dev.sergiobelda.todometer.common.ui.error
 
 import dev.sergiobelda.todometer.common.domain.Result
 
@@ -22,7 +22,7 @@ data class ErrorUi(
     val message: String? = null
 )
 
-internal fun <T : Any> Result<T>.mapToErrorUi(): ErrorUi? =
+fun <T : Any> Result<T>.mapToErrorUi(): ErrorUi? =
     if (this is Result.Error) {
         ErrorUi(message = message)
     } else null
